+++ server/src/internalClusterTest/java/org/opensearch/snapshots/RestoreSnapshotIT.java
@@ -87,7 +87,7 @@ import static org.hamcrest.Matchers.nullValue;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_REPLICAS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_SHARDS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REMOTE_STORE_ENABLED;
-import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REMOTE_STORE_REPOSITORY;
+import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY;
 import static org.opensearch.index.IndexSettings.INDEX_REFRESH_INTERVAL_SETTING;
 import static org.opensearch.index.IndexSettings.INDEX_SOFT_DELETES_SETTING;
 import static org.opensearch.index.query.QueryBuilders.matchQuery;
@@ -304,7 +304,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
             .get();
         indexSettings = getIndexResponse.settings().get(restoredIndexName1Seg);
         assertNull(indexSettings.get(SETTING_REMOTE_STORE_ENABLED));
-        assertNull(indexSettings.get(SETTING_REMOTE_STORE_REPOSITORY, null));
+        assertNull(indexSettings.get(SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, null));
         assertEquals(ReplicationType.SEGMENT.toString(), indexSettings.get(IndexMetadata.SETTING_REPLICATION_TYPE));
         assertDocsPresentInIndex(client, restoredIndexName1Seg, numDocsInIndex1);
         // indexing some new docs and validating
@@ -331,7 +331,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
             .get();
         indexSettings = getIndexResponse.settings().get(restoredIndexName1Doc);
         assertNull(indexSettings.get(SETTING_REMOTE_STORE_ENABLED));
-        assertNull(indexSettings.get(SETTING_REMOTE_STORE_REPOSITORY, null));
+        assertNull(indexSettings.get(SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, null));
         assertNull(indexSettings.get(IndexMetadata.SETTING_REPLICATION_TYPE));
         assertDocsPresentInIndex(client, restoredIndexName1Doc, numDocsInIndex1);
         // indexing some new docs and validating
@@ -492,7 +492,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
         assertThat(createSnapshotResponse.getSnapshotInfo().state(), equalTo(SnapshotState.SUCCESS));
 
         Settings remoteStoreIndexSettings = Settings.builder()
-            .put(IndexMetadata.SETTING_REMOTE_STORE_REPOSITORY, remoteStoreRepo2Name)
+            .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, remoteStoreRepo2Name)
             .build();
         // restore index as a remote store index with different remote store repo
         RestoreSnapshotResponse restoreSnapshotResponse = client.admin()
@@ -532,7 +532,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
             .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, numOfReplicas);
         if (enableRemoteStore) {
             settingsBuilder.put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, true)
-                .put(IndexMetadata.SETTING_REMOTE_STORE_REPOSITORY, remoteStoreRepo)
+                .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, remoteStoreRepo)
                 .put(IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY, remoteStoreRepo)
                 .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), "300s")
                 .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT);
@@ -569,7 +569,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
         Settings indexSettings = Settings.builder()
             .put(super.indexSettings())
             .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, true)
-            .put(IndexMetadata.SETTING_REMOTE_STORE_REPOSITORY, remoteStoreRepoName)
+            .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, remoteStoreRepoName)
             .put(IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY, remoteStoreRepoName)
             .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), "300s")
             .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)