+++ server/src/test/java/org/opensearch/index/IndexSettingsTests.java
@@ -54,6 +54,7 @@ import org.opensearch.test.VersionUtils;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.Locale;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -846,7 +847,7 @@ public class IndexSettingsTests extends OpenSearchTestCase {
             Settings.builder()
                 .put(IndexMetadata.SETTING_VERSION_CREATED, Version.CURRENT)
                 .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, true)
-                .put(IndexMetadata.SETTING_REMOTE_STORE_REPOSITORY, "repo1")
+                .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, "repo1")
                 .build()
         );
         IndexSettings settings = new IndexSettings(metadata, Settings.EMPTY);
@@ -860,42 +861,57 @@ public class IndexSettingsTests extends OpenSearchTestCase {
         SettingsException error = expectThrows(
             SettingsException.class,
             () -> settings.updateSettings(
-                Settings.builder().put("index.remote_store.repository", randomUnicodeOfLength(10)).build(),
+                Settings.builder().put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, randomUnicodeOfLength(10)).build(),
                 Settings.builder(),
                 Settings.builder(),
                 "index"
             )
         );
-        assertEquals(error.getMessage(), "final index setting [index.remote_store.repository], not updateable");
+        assertEquals(
+            error.getMessage(),
+            String.format(Locale.ROOT, "final index setting [%s], not updateable", IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY)
+        );
     }
 
     public void testSetRemoteRepositoryFailsWhenRemoteStoreIsNotEnabled() {
         Settings indexSettings = Settings.builder()
-            .put("index.replication.type", ReplicationType.SEGMENT)
-            .put("index.remote_store.enabled", false)
-            .put("index.remote_store.repository", "repo1")
+            .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)
+            .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, false)
+            .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, "repo1")
             .build();
         IllegalArgumentException iae = expectThrows(
             IllegalArgumentException.class,
             () -> IndexMetadata.INDEX_REMOTE_STORE_REPOSITORY_SETTING.get(indexSettings)
         );
         assertEquals(
-            "Settings index.remote_store.repository can only be set/enabled when index.remote_store.enabled is set to true",
+            String.format(
+                Locale.ROOT,
+                "Settings %s can only be set/enabled when %s is set to true",
+                IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY,
+                IndexMetadata.SETTING_REMOTE_STORE_ENABLED
+            ),
             iae.getMessage()
         );
     }
 
     public void testSetRemoteRepositoryFailsWhenEmptyString() {
         Settings indexSettings = Settings.builder()
-            .put("index.replication.type", ReplicationType.SEGMENT)
-            .put("index.remote_store.enabled", false)
-            .put("index.remote_store.repository", "")
+            .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)
+            .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, false)
+            .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, "")
             .build();
         IllegalArgumentException iae = expectThrows(
             IllegalArgumentException.class,
             () -> IndexMetadata.INDEX_REMOTE_STORE_REPOSITORY_SETTING.get(indexSettings)
         );
-        assertEquals("Setting index.remote_store.repository should be provided with non-empty repository ID", iae.getMessage());
+        assertEquals(
+            String.format(
+                Locale.ROOT,
+                "Setting %s should be provided with non-empty repository ID",
+                IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY
+            ),
+            iae.getMessage()
+        );
     }
 
     public void testRemoteTranslogRepoDefaultSetting() {
@@ -924,7 +940,7 @@ public class IndexSettingsTests extends OpenSearchTestCase {
 
     public void testSetRemoteTranslogRepositoryFailsWhenRemoteStoreIsNotEnabled() {
         Settings indexSettings = Settings.builder()
-            .put("index.replication.type", ReplicationType.SEGMENT)
+            .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)
             .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, false)
             .put(IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY, "repo1")
             .build();
@@ -940,7 +956,7 @@ public class IndexSettingsTests extends OpenSearchTestCase {
 
     public void testSetRemoteTranslogRepositoryFailsWhenEmptyString() {
         Settings indexSettings = Settings.builder()
-            .put("index.replication.type", ReplicationType.SEGMENT)
+            .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)
             .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, true)
             .put(IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY, "")
             .build();